feat: unify route and location verification into single admin view, closes #54
Replace separate RouteVerificationView and LocationVerificationView with a
unified VerificationView at /admin/verify. Groups pending submissions into
collapsible Route and Location sections with count badges. Inline actions
(approve, edit, merge, reject for routes; approve, reject for locations)
work without leaving the view. Shows empty state when nothing is pending.
